should not delete an user if he has an transaction

+++ app/Http/Controllers/Api/UserController.php
@@ -38,6 +38,9 @@ class UserController extends Controller
 
     public function destroy(User $user): JsonResponse
     {
+        if ($user->wallet->transactions()->count() > 0) {
+            return response()->json(['message' => 'User cannot be deleted because it has transactions'], 403);
+        }
         $user->delete();
         return response()->json(['message' => 'User deleted successfully'], 200);
     }
